Change AST transforms to mutate the original AST
Creating a new one has the issue that variables are still shared. Variables
contain information such as whether it's ever set! or whether it's initialised
to a constant, which is something that can change with AST transforms.

Unsharing the variables then has the issue that top-level values are uniquely
identified by the variables. So now either the top-level values have to be
duplicated as well, affecting program semantics, or top-level variables have to
remain shared which still has the problem above.

Overall, it seems easier to just transform the AST in an imperative manner
rather than a functional one.